Move some code into util/blobs/
