+++ packages/beacon-node/src/api/impl/beacon/blocks/index.ts
@@ -214,6 +214,7 @@ export function getBeaconBlockApi({
 
       console.log("BLOB SAVING 1. publishBlockWithBlobs is calling chain.processBlock");
 
+      // TODO EIP-4844 is it appropriate to do this here?
       await db.blob.add(blobsSidecar);
 
       console.log("Persisted blobsSidecar to the database");
